Move filter logic to be single-resource based in buildBundle.ts
